Modularize wv.ui.js
